+++ odoo/local-src/base_stock_picking_pack_split/__openerp__.py
@@ -1,23 +1,6 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Yannick Vaucher
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2013-2016 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 {'name': 'Base stock picking pack split',
  'version': '1.0',
  'category': 'Warehouse',
@@ -44,11 +27,11 @@ Contributors
  'author': 'Camptocamp',
  'maintainer': 'Camptocamp',
  'website': 'http://www.camptocamp.com/',
- 'depends': ['stock'],
+ 'depends': ['stock', 'delivery'],  # delivery not needed, only to pass tests
  'data': ['wizard/split_packs_view.xml',
           ],
  'test': [],
- 'installable': False,
+ 'installable': True,
  'auto_install': False,
  'application': True,
  }