+++ odoo/local-src/delivery_carrier_label_postlogistics_shop_logo/models/__init__.py
@@ -1,0 +1,2 @@
+from . import qoqa_shop
+from . import stock