+++ odoo/local-src/delivery_carrier_label_swiss_pp/__init__.py
@@ -1,7 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import company
-from . import sale
-from . import res_config
-from . import report
-from . import delivery
-from . import stock
+from . import models