+++ HISTORY.rst
@@ -29,6 +29,10 @@ latest (unreleased)
 * Validating invoices takes less time.
 * Creating an invoice from a SO takes less time.
 * Migrate Wine report addon
+* Migrate addon to add a wizard to split products in multiples packs
+* Migrate Swiss PP labels addon
+* Migrate addon to select a logo per shop on postlogistics delivery labels
+* Migrate Swiss PP labels addon
 
 **Bugfixes**
 * Analytic accounts : allow to "search more..." on SO